Daily Quota 3B
I might aswell use this space up but I'm unsure what to write here so that it provides useful context so I'm gonna use this as as personal diary of sorts.

Hotline miami 2 : wrong number is a REALLY good game the lore goes really deep and its easy to miss details and the game itself is pretty fun I like it more than the first honestly.