Daily Quota 16A
Do you see this? ... Can you see my face? This is my true nature! You see, don't you? This is who I am! This is who we all are. We're animals! ... There's no denying it! A bunch of goddamn animals! They're sending us out to slaughter or be slaughtered... And here we sit until they tell us what to do, and how to do it! No will of our own. Just mindless obedience! We don't even know why we're fighting now, do we? All we know is that deep down, somewhere in there, we enjoy it. Destruction and violence... It's just part of our nature.